Move contextswitcher open state to Jotai
