+++ src/redux/action-types.ts
@@ -15,8 +15,6 @@ export const GLOBAL_FILTER_REMOVE = '@@chrome/global-filter-remove';
 
 export const REGISTER_MODULE = '@@chrome/register-async-module';
 
-export const TOGGLECONTEXTSWITCHER = '@@chrome/context-switcher-toggle';
-
 export const LOAD_NAVIGATION_LANDING_PAGE = '@@chrome/load-navigation-landing-page';
 export const LOAD_LEFT_NAVIGATION_SEGMENT = '@@chrome/load-navigation-segment';
 export const LOAD_MODULES_SCHEMA = '@@chrome/load-modules-schema';
